TestScaffold: startTo: suspend running threads if debugging is started delayed. (#4)
TestScaffold.startTo() has a race codition when debugging is started on demand, because at that
point all debuggee threads are running when resumeTo() is called. The race is between suspend upon
hitting the breakpoint and the resume after setting the breakpoint. If the resume occurs after the
suspend we get an com.sun.jdi.IncompatibleThreadStateException when trying to retrieve the frames in
ListenerTest.runTests().